refactor: messages module
With so many files that were somewhat the same, I aimed to orgainize the repository better to make functions more clear. This is because the remaining top level files (const.rs, parser.rs) act on and consume all of the files and structs within `messages`.
