+++ src/messages/mod.rs
@@ -1,0 +1,8 @@
+pub mod activate_ext;
+pub mod error_message;
+pub mod get_shares;
+pub mod get_window;
+pub mod new_block_found;
+pub mod new_txs;
+pub mod share_ok;
+pub mod verify_fees;
